+++ docs/conf.py
@@ -25,7 +25,7 @@ project = 'conmech'
 year = '2019'
 author = 'Yijiang Huang'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '0.3.0'
+version = release = '0.3.1'
 
 pygments_style = 'trac'  # Perhaps change to sphinx
 templates_path = ['.']